The following - possibly updated - changelog can be viewed as formatted
article at https://phabricator.mitk.org/w/mitk/changelog/2022.17/.

NOTE: This is the last snapshot before the next official MITK release.

= 🛠 Third-party dependency changes =

//none//

= ✨ New features =

- Segmentation
 - Improved default color palette for labels
 - Improved generic naming and coloring of newly created labels
 - Added preference (and command-line argument `--Segmentation.labelSetPreset`) for default label set presets
- Volume Visualization
 - Simplified the interface of volume visualization

= 🐛 Bugfixes =

- Segmentation
 - Made some segmentation tools more aware of the active label and labels in general
 - Fixed invisible contours in some cases when drawing in multiple render windows
 - Fixed crash when creating polygon models of segmentations (was recently introduced with ITK v5 upgrade)
 - Fixed crash when using the picking tool on dynamic segmentations
 - Fixed selection of multiple labels
 - Removed actions from labels context menu that never were implemented
 - Removed hard to use fast marching and watershed segmentation tools
 - Various fixes and improvements in the experimental nnU-Net segmentation tool
- Image Cropper
 - Fixed repetition in naming of cropped images
- Preferences
 - Fixed default render window colors when resetting preferences for Standard Multi Widget

= 🔥 API-breaking changes =

=== Volume Visualization ===

The simplification of the Volume Visualization UI allowed us to get rid of otherwise unused legacy classes and properties.

The following classes were removed without replacement;

- 📂 Modules
 - 📂 Core
  - VtkVolumeRenderingProperty
 - 📂 MapperExt
  - GPUVolumeMapper3D
  - vtkMitkGPUVloumeRayCastMapper
  - vtkMitkOpenGLVolumeTextureMapper3D
  - vtkMitkVolumeTextureMapper3D

The following properties are no longer used by MITK or were renamed:

- volumerendering.cpu.*
- volumerendering.ray.*
- volumerendering.gpu.reducesliceartifacts
- volumerendering.gpu.usetexturecompression
- volumerendering.gpu.* {icon arrow-right} volumerendering.*
- volumerendering.usegpu
- volumerendering.uselod
- volumerendering.usemip
- volumerendering.useray

+++ Modules/IGT/TrackingDevices/mitkClaronInterface.h
@@ -24,7 +24,7 @@ found in the LICENSE file.
 #include <itkObjectFactory.h>
 
 #ifdef _WIN64  //Defined for applications for Win64.
-typedef long mtHandle;
+typedef long long mtHandle;
 #else
 typedef int mtHandle;
 #endif
